+++ src/main/java/com/github/davidepastore/jbugmenot/Account.java
@@ -3,6 +3,8 @@
  */
 package com.github.davidepastore.jbugmenot;
 
+import java.io.IOException;
+
 /**
  * An account entity.
  * 
@@ -17,6 +19,7 @@ public class Account {
 	private String stats;
 	private long votes;
 	private long id;
+	private long site;
 
 	/**
 	 * @return the username
@@ -106,6 +109,29 @@ public class Account {
 	 */
 	public void setId(long id) {
 		this.id = id;
+	}
+
+	/**
+	 * @return the site
+	 */
+	public long getSite() {
+		return site;
+	}
+
+	/**
+	 * @param site the site to set
+	 */
+	public void setSite(long site) {
+		this.site = site;
+	}
+	
+	/**
+	 * Vote this account using a positive or negative vote.
+	 * @param vote The vote. True to mark this as good login, false otherwise.
+	 * @throws IOException 
+	 */
+	public void vote(boolean vote) throws IOException{
+		JBugmenot.vote(this, vote);
 	}
 
 }